+++ ListableUI/Tests/Item/ItemContentCoordinatorTests.swift
@@ -23,23 +23,17 @@ class ItemContentCoordinatorActionsTests : XCTestCase
             callbackCount += 1
         })
         
-        self.testcase("Setter based update") {
-        
-            var updated = item
-            updated.content.value = "update1"
-            actions.update(updated)
-            
-            XCTAssertEqual(item.content.value, "update1")
-            XCTAssertEqual(callbackCount, 1)
-        }
-        
         self.testcase("Closure based update") {
                         
             actions.update {
                 $0.content.value = "update2"
             }
             
-            XCTAssertEqual(item.content.value, "update2")
+            actions.update {
+                $0.content.value = "update3"
+            }
+            
+            XCTAssertEqual(item.content.value, "update3")
             XCTAssertEqual(callbackCount, 2)
         }
     }